Moving everything to main directory
