+++ kicks.py
