+++ square.py
